requirements.txt
streamlit==1.28.0
streamlit-extras==0.3.5
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.0.1